implement parsing of subprocess
