+++ src/new_model/parser/index.ts
@@ -1,3 +1,6 @@
 export * from './activity_parser';
 export * from './event_parser';
 export * from './gateway_parser';
+export * from './process_parser';
+export * from './process_lane_set_parser';
+export * from './collaboration_parser';
